rename: briefStationInfo를 stationSummary로 이름 변경
[#382]
